+++ util/modelUtil.py
@@ -5,6 +5,12 @@ import os
 import numpy as np
 import matplotlib.pyplot as plt
 from sklearn.model_selection import train_test_split
+from sklearn.svm import SVC
+from sklearn.svm import LinearSVC
+from sklearn.pipeline import make_pipeline
+from sklearn.preprocessing import StandardScaler
+from sklearn.datasets import make_classification
+from scipy.linalg import sqrtm
 from util.nets import *
 import json
 import mne
@@ -184,3 +190,74 @@ def SVM(X_train, y_train, X_test, y_test):
             counter += 1
     acc = counter * 100.0 / len(predicted)
     return acc
+
+def PSD(X_train, y_train, X_test, y_test):
+    tmp = []
+    for xxx in X_train:
+        xx = xxx.T
+        fft_rs, freq = GetFFT(xx)
+        newX = GetPSD(fft_rs)
+        newX = np.asarray(newX)
+        newX = newX.real
+        newX = newX.reshape(-1)
+        tmp.append(newX)
+    X_train = np.vstack(tmp)
+
+    tmp = []
+    for xxx in X_test:
+        xx = xxx.T
+        fft_rs, freq = GetFFT(xx)
+        newX = GetPSD(fft_rs)
+        newX = np.asarray(newX)
+        newX = newX.real
+        newX = newX.reshape(-1)
+        tmp.append(newX)
+    X_test = np.vstack(tmp)
+    return SVM(X_train, y_train, X_test, y_test)
+
+
+def IHAR(X_train, y_train, X_test, y_test, listChns):
+    
+    electrodeIHAR = [['Fp1', 'F7', 'F3', 'FC5', 'T7', 'P7', 'O1'], ['Fp2', 'F8', 'F4', 'FC6', 'T8', 'P8', 'O2']]
+    numNode = len(electrodeIHAR[0])
+    electrodeIndex = []
+    for electrodes in electrodeIHAR:
+        electrodeIndex.append([listChns.index(x) for x in electrodes])
+    tmp = []
+    for xxx in X_train:
+        xx = xxx.T
+        fft_rs, freq = GetFFT(xx)
+        newX = GetPSD(fft_rs)
+        newX = np.asarray(newX)
+        newX = newX.real
+        newX = MA(newX, args.windowIHAR)
+        IharTrain = []
+        for ii in range(numNode):
+            left = newX[electrodeIndex[0][ii]]
+            right = newX[electrodeIndex[1][ii]]
+            ihar = left / right
+            IharTrain.append(ihar)
+        newX = np.vstack([np.vstack(IharTrain), newX])
+        newX = newX.reshape(-1)
+        tmp.append(newX)
+    X_train = np.vstack(tmp)
+
+    tmp = []
+    for xxx in X_test:
+        xx = xxx.T
+        fft_rs, freq = GetFFT(xx)
+        newX = GetPSD(fft_rs)
+        newX = np.asarray(newX)
+        newX = newX.real
+        newX = MA(newX, args.windowIHAR)
+        IharTrain = []
+        for ii in range(numNode):
+            left = newX[electrodeIndex[0][ii]]
+            right = newX[electrodeIndex[1][ii]]
+            ihar = left / right
+            IharTrain.append(ihar)
+        newX = np.vstack([np.vstack(IharTrain), newX])
+        newX = newX.reshape(-1)
+        tmp.append(newX)
+    X_test = np.vstack(tmp)    
+    return SVM(X_train, y_train, X_test, y_test)
